Create basic controllers and views
Note that this was walking along with a lesson that was not intended
to be used in a functional app. I was trying to jury rig it without
the experience needed to get it all up and running. Moving on with
the reading to learn more. Next project will be more complete.